登录+tabbar

+++ YuYi_Server/AppDelegate.m
@@ -72,7 +72,7 @@
             NSLog(@"登陆成功。当前登录的用户ID：%@", userId);
             [[RCIM sharedRCIM] setReceiveMessageDelegate:self];
         } error:^(RCConnectErrorCode status) {
-            NSLog(@"登陆的错误码为:%d", status);
+            NSLog(@"登陆的错误码为:%ld", (long)status);
         } tokenIncorrect:^{
             //token过期或者不正确。
             //如果设置了token有效期并且token过期，请重新请求您的服务器获取新的token
